+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 3.4.0-x
+
+- **Breaking change** The `derive.chain.{getBlock, subscribeNewBlocks}` now return `SignedBlockExtended`, all with the actual extrinsics and events mapped. Users of the latter interface may need some updates since the result structure may be different.
+
+Changes:
+
+- `derive.chain.getBlock()` now maps events to extrinsics via `.extrinsics` getter
+
+
 ## 3.3.1 Jan 4, 2021
 
 Upgrade priority: Low. Recommended for users of frontier evm module with additional types.